+++ client/js/src/test/data.test.ts
@@ -300,12 +300,7 @@ describe("handle_file", () => {
 		}
 		const file = new File(["test image"], "test.png", { type: "image/png" });
 		const result = handle_file(file) as FileData;
-		expect(result.path).toBe("test.png");
-		expect(result.orig_name).toBe("test.png");
-		expect(result.blob).toBeInstanceOf(Blob);
-		expect(result.size).toBe(file.size);
-		expect(result.mime_type).toBe("image/png");
-		expect(result.meta).toEqual({ _type: "gradio.FileData" });
+		expect(result).toBeInstanceOf(Blob);
 	});
 
 	it("should throw an error for invalid input", () => {
@@ -367,7 +362,7 @@ describe("handle_payload", () => {
 	it("should return an output payload without the state component value when with_null_state is false", () => {
 		const resolved_payload = ["hello", null];
 		const dependency = {
-			inputs: [2, 3]
+			outputs: [2, 3]
 		};
 		const components = [
 			{ id: 2, type: "textbox" },
@@ -388,7 +383,7 @@ describe("handle_payload", () => {
 	it("should return an ouput payload without the two state component values when with_null_state is false", () => {
 		const resolved_payload = ["hello", null, "world", null];
 		const dependency = {
-			inputs: [2, 3, 4, 5]
+			outputs: [2, 3, 4, 5]
 		};
 		const components = [
 			{ id: 2, type: "textbox" },
@@ -411,7 +406,7 @@ describe("handle_payload", () => {
 	it("should return an ouput payload with the two state component values when with_null_state is true", () => {
 		const resolved_payload = ["hello", null, "world", null];
 		const dependency = {
-			inputs: [2, 3, 4, 5]
+			outputs: [2, 3, 4, 5]
 		};
 		const components = [
 			{ id: 2, type: "textbox" },
@@ -446,6 +441,7 @@ describe("handle_payload", () => {
 			// @ts-ignore
 			dependency,
 			components,
+			"input",
 			with_null_state
 		);
 		expect(result).toEqual(["hello", "world"]);