+++ gradio/cli/commands/components/_create_utils.py
@@ -204,6 +204,16 @@ OVERRIDES = {
         python_file_name="image_editor.py",
         js_dir="imageeditor",
     ),
+    "MultimodalTextbox": ComponentFiles(
+        template="MultimodalTextbox",
+        python_file_name="multimodal_textbox.py",
+        js_dir="multimodaltextbox",
+    ),
+    "DownloadButton": ComponentFiles(
+        template="DownloadButton",
+        python_file_name="download_button.py",
+        js_dir="downloadbutton",
+    ),
 }