+++ js/markdown/shared/MarkdownCode.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import { afterUpdate, createEventDispatcher } from "svelte";
+	import { afterUpdate } from "svelte";
 	import DOMPurify from "dompurify";
 	import render_math_in_element from "katex/contrib/auto-render";
 	import "katex/dist/katex.min.css";
@@ -73,15 +73,18 @@
 			}
 		}
 	}
-	afterUpdate(() => render_html(message));
+
+	afterUpdate(async () => {
+		if (el && document.body.contains(el)) {
+			await render_html(message);
+		} else {
+			console.error("Element is not in the DOM");
+		}
+	});
 </script>
 
 <span class:chatbot bind:this={el} class="md" class:prose={render_markdown}>
-	{#if render_markdown}
-		{@html html}
-	{:else}
-		{html}
-	{/if}
+	{@html html}
 </span>
 
 <style>